+++ es5/index.js
@@ -1,0 +1,51 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.FluxComponent = exports.Joins = exports.Store = exports.Publisher = exports.SyncAction = exports.PromiseAction = exports.AsyncResultAction = exports.Action = undefined;
+
+var _Action2 = require('./Action');
+
+var _Action3 = _interopRequireDefault(_Action2);
+
+var _AsyncResultAction2 = require('./AsyncResultAction');
+
+var _AsyncResultAction3 = _interopRequireDefault(_AsyncResultAction2);
+
+var _PromiseAction2 = require('./PromiseAction');
+
+var _PromiseAction3 = _interopRequireDefault(_PromiseAction2);
+
+var _SyncAction2 = require('./SyncAction');
+
+var _SyncAction3 = _interopRequireDefault(_SyncAction2);
+
+var _Publisher2 = require('./Publisher');
+
+var _Publisher3 = _interopRequireDefault(_Publisher2);
+
+var _Store2 = require('./Store');
+
+var _Store3 = _interopRequireDefault(_Store2);
+
+var _JoinStores = require('./JoinStores');
+
+var _Joins = _interopRequireWildcard(_JoinStores);
+
+var _FluxComponent2 = require('./FluxComponent');
+
+var _FluxComponent3 = _interopRequireDefault(_FluxComponent2);
+
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+exports.Action = _Action3.default;
+exports.AsyncResultAction = _AsyncResultAction3.default;
+exports.PromiseAction = _PromiseAction3.default;
+exports.SyncAction = _SyncAction3.default;
+exports.Publisher = _Publisher3.default;
+exports.Store = _Store3.default;
+exports.Joins = _Joins;
+exports.FluxComponent = _FluxComponent3.default;